[CC1101] Changed pin mapping

+++ examples/SX1231/SX1231_Receive/SX1231_Receive.ino
@@ -15,9 +15,9 @@
 #include <RadioLib.h>
 
 // SX1231 has the following connections:
-// NSS pin:   10
+// CS pin:    10
 // DIO0 pin:  2
-// DIO1 pin:  3
+// RESET pin: 3
 SX1231 rf = new Module(10, 2, 3);
 
 // or using RadioShield
@@ -35,7 +35,7 @@ void setup() {
   // frequency deviation:                 50.0 kHz
   // output power:                        13 dBm
   // sync word:                           0x2D01
-  byte state = rf.begin();
+  int state = rf.begin();
   if (state == ERR_NONE) {
     Serial.println(F("success!"));
   } else {
